api/requirements.txt
Flask-SQLAlchemy
wtforms_alchemy
pytest-flask
Flask-WTF
gunicorn
flask-restplus
pep8
passlib
marshmallow
flask_redis